replace random with stdlib rand

+++ game/core/game_host.go
@@ -90,11 +90,11 @@ func MakeHost(setup Setup, options HostOptions) *GameHost {
 		for j := range players {
 			ctrl := f.Ctrl(byte(j))
 
-			ctrl.PlayerIndex = players[j].Index
+			ctrl.PlayerIndex = byte(players[j].Index)
 			ctrl.Name = players[j].Name
 			ctrl.Feed = piece.NewCtrlFeed(setup.Config.PieceFeed, i, j, setup.Config.SamePieces)
 			ctrl.Config = players[j].Config
-			ctrl.SetLevel(setup.Config.Level)
+			ctrl.SetLevel(uint(setup.Config.Level))
 			ctrl.IsShadowShown = true
 
 			ctrl.IsColumnLimited = setup.Config.PlayerZones
