+++ game/core/game_iterpreter.go
@@ -76,11 +76,11 @@ func MakeInterpreter(setup Setup, options InterpreterOptions) *GameInterpreter {
 		for j := range players {
 			ctrl := f.Ctrl(byte(j))
 
-			ctrl.PlayerIndex = players[j].Index
+			ctrl.PlayerIndex = byte(players[j].Index)
 			ctrl.Name = players[j].Name
 			ctrl.Feed = piece.NewCtrlFeed(setup.Config.PieceFeed, i, j, setup.Config.SamePieces)
 			ctrl.Config = players[j].Config
-			ctrl.SetLevel(setup.Config.Level)
+			ctrl.SetLevel(uint(setup.Config.Level))
 			ctrl.IsShadowShown = true
 
 			ctrl.IsColumnLimited = setup.Config.PlayerZones
